Portal Behaviour Fixes V2
Fixes #118

- fixes the falling sound previously heard when standing in between two portals (used actual velocity of body moving)
- fixes the player falling out of the world if they somehow get behind a portal (just teleports to other portal if behind)
- fixes the clipping issue when a player shoots the color portal they are currently standing in
	this is done by "nudging" the player out of the portal slightly before it is closed, their velocity is also set to 0vec in this edge case.
- same behaviour outlined in point above if same situations happens from a cutscene (player is nudged)

I have confirmed that bullet point 2 now allows the player to clip into a portal from behOind it in test chamber 00. I think this should be resolved by adding some depth to that wall behind the portal(change the geometry of that level a bit).
I can do that in a follow-up commit. to my knowledge there arent any other instances where the player should be able to get to a spot right behind an active portal.